requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.4.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
Pillow==7.2.0
PyJWT==1.7.1
pymongo==3.10.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wincertstore==0.2
